+++ sc-carbon-footprint/src/intf/Constants.sol
@@ -2,7 +2,7 @@
 
 pragma solidity 0.8.17;
 
-uint constant Const_BlockDelayBeforeVote = uint(keccak256("BlockDelayBeforeVote"));
-uint constant Const_BlockSpanForVote = uint(keccak256("BlockSpanForVote"));
-uint constant Const_MaxNbBlockPerPeriod = uint(keccak256("MaxNbBlockPerPeriod"));
-uint constant Const_MinPledgeAmountWei = uint(keccak256("MinPledgeAmountWei"));
+uint256 constant Const_BlockDelayBeforeVote = uint256(keccak256("BlockDelayBeforeVote"));
+uint256 constant Const_BlockSpanForVote = uint256(keccak256("BlockSpanForVote"));
+uint256 constant Const_MaxNbBlockPerPeriod = uint256(keccak256("MaxNbBlockPerPeriod"));
+uint256 constant Const_MinPledgeAmountWei = uint256(keccak256("MinPledgeAmountWei"));
