+++ sc-carbon-footprint/src/intf/ICarbonFootprint.sol
@@ -3,9 +3,9 @@
 pragma solidity 0.8.17;
 
 interface ICarbonFootprint {
-  function setFootprint(address node, uint value) external; 
-  function nbNodes() external view returns (uint);
-  function totalFootprint() external view returns (uint);
-  function footprint(address node) external view returns (uint);
-  event CarbonFootprintUpdate(address indexed node, uint footprint);
+  function setFootprint(address node, uint256 value) external; 
+  function nbNodes() external view returns (uint256);
+  function totalFootprint() external view returns (uint256);
+  function footprint(address node) external view returns (uint256);
+  event CarbonFootprintUpdate(address indexed node, uint256 footprint);
 }
